+++ event.go
@@ -1,0 +1,97 @@
+// Copyright 2016 The go-libvirt Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Package libvirt is a pure Go implementation of the libvirt RPC protocol.
+// For more information on the protocol, see https://libvirt.org/internals/l.html
+package libvirt
+
+import (
+	"bytes"
+	"errors"
+
+	"github.com/davecgh/go-xdr/xdr2"
+	"github.com/digitalocean/go-libvirt/internal/constants"
+)
+
+// ErrEventsNotSupported is returned by Events() if event streams
+// are unsupported by either QEMU or libvirt.
+var ErrEventsNotSupported = errors.New("event monitor is not supported")
+
+// DomainEvent represents a libvirt domain event.
+type DomainEvent struct {
+	CallbackID   uint32
+	Domain       Domain
+	Event        string
+	Seconds      uint64
+	Microseconds uint32
+	Padding      uint8
+	Details      []byte
+}
+
+// DomainEvents streams domain events.
+// If a problem is encountered setting up the event monitor connection
+// an error will be returned. Errors encountered during streaming will
+// cause the returned event channel to be closed.
+func (l *Libvirt) DomainEvents(d *Domain) (<-chan DomainEvent, error) {
+	payload := struct {
+		Padding [4]byte
+		Domain  Domain
+		Event   [2]byte
+		Flags   [2]byte
+	}{
+		Padding: [4]byte{0x0, 0x0, 0x1, 0x0},
+		Domain:  *d,
+		Event:   [2]byte{0x0, 0x0},
+		Flags:   [2]byte{0x0, 0x0},
+	}
+
+	buf, err := encode(&payload)
+	if err != nil {
+		return nil, err
+	}
+
+	resp, err := l.request(constants.QEMUConnectDomainMonitorEventRegister, constants.ProgramQEMU, &buf)
+	if err != nil {
+		return nil, err
+	}
+
+	res := <-resp
+	if res.Status != StatusOK {
+		err := decodeError(res.Payload)
+		if err == ErrUnsupported {
+			return nil, ErrEventsNotSupported
+		}
+
+		return nil, decodeError(res.Payload)
+	}
+
+	dec := xdr.NewDecoder(bytes.NewReader(res.Payload))
+
+	cbID, _, err := dec.DecodeUint()
+	if err != nil {
+		return nil, err
+	}
+
+	stream := make(chan *DomainEvent)
+	l.addStream(cbID, stream)
+	c := make(chan DomainEvent)
+	go func() {
+		// process events
+		for e := range stream {
+			c <- *e
+		}
+	}()
+
+	return c, nil
+}
